chore: Move number badge to jobs management
