+++ web_ui/src/shared/components/header/jobs-management/number-badge/number-badge.module.scss
